Move Verifier handling to a trait
